refactor: split entities table generate CSV file into hook
