+++ apps/hash-frontend/src/pages/shared/entities-table.tsx
@@ -1,9 +1,6 @@
-import { useLazyQuery } from "@apollo/client";
 import {
   Entity as BpEntity,
   EntityRootType as BpEntityRootType,
-  EntityType,
-  PropertyType,
   Subgraph as BpSubgraph,
 } from "@blockprotocol/graph";
 import {
@@ -14,32 +11,16 @@ import {
 } from "@glideapps/glide-data-grid";
 import { EntitiesGraphChart } from "@hashintel/block-design-system";
 import { ListRegularIcon } from "@hashintel/design-system";
-import { generateEntityLabel } from "@local/hash-isomorphic-utils/generate-entity-label";
-import {
-  currentTimeInstantTemporalAxes,
-  mapGqlSubgraphFieldsFragmentToSubgraph,
-  zeroedGraphResolveDepths,
-} from "@local/hash-isomorphic-utils/graph-queries";
 import { isPageEntityTypeId } from "@local/hash-isomorphic-utils/page-entity-type-ids";
 import { simplifyProperties } from "@local/hash-isomorphic-utils/simplify-properties";
 import { PageProperties } from "@local/hash-isomorphic-utils/system-types/shared";
 import {
   Entity,
   EntityId,
-  EntityRootType,
-  EntityTypeWithMetadata,
   extractEntityUuidFromEntityId,
   extractOwnedByIdFromEntityId,
 } from "@local/hash-subgraph";
-import {
-  getEntityTypeById,
-  getRightEntityForLinkEntity,
-  getRoots,
-} from "@local/hash-subgraph/stdlib";
-import {
-  extractBaseUrl,
-  LinkEntity,
-} from "@local/hash-subgraph/type-system-patch";
+import { extractBaseUrl } from "@local/hash-subgraph/type-system-patch";
 import {
   Box,
   ToggleButton,
@@ -67,12 +48,6 @@ import {
 import { BlankCell, blankCell } from "../../components/grid/utils";
 import { ColumnFilter } from "../../components/grid/utils/filtering";
 import { useGetOwnerForEntity } from "../../components/hooks/use-get-owner-for-entity";
-import {
-  StructuralQueryEntitiesQuery,
-  StructuralQueryEntitiesQueryVariables,
-} from "../../graphql/api-types.gen";
-import { structuralQueryEntitiesQuery } from "../../graphql/queries/knowledge/entity.queries";
-import { MinimalUser } from "../../lib/user-and-org";
 import { useEntityTypeEntitiesContext } from "../../shared/entity-type-entities-context";
 import { ChartNetworkRegularIcon } from "../../shared/icons/chart-network-regular-icon";
 import { HEADER_HEIGHT } from "../../shared/layout/layout-with-header/page-header";
@@ -92,6 +67,7 @@ import {
   TypeEntitiesRow,
   useEntitiesTable,
 } from "./entities-table/use-entities-table";
+import { useEntitiesTableGenerateCsvFile } from "./entities-table/use-entities-table-generate-csv-file";
 import { TOP_CONTEXT_BAR_HEIGHT } from "./top-context-bar";
 
 export const EntitiesTable: FunctionComponent<{
@@ -465,244 +441,13 @@ export const EntitiesTable: FunctionComponent<{
     ],
   );
 
-  const [structuralQueryEntities] = useLazyQuery<
-    StructuralQueryEntitiesQuery,
-    StructuralQueryEntitiesQueryVariables
-  >(structuralQueryEntitiesQuery);
-
-  const fetchOutgoingLinksOfEntities = useCallback(
-    async (params: {
-      leftEntities: Entity[];
-    }): Promise<
-      {
-        linkEntity: LinkEntity;
-        rightEntity: Entity;
-        rightEntityLabel: string;
-        linkEntityType: EntityTypeWithMetadata;
-      }[]
-    > => {
-      const { leftEntities } = params;
-
-      const { data } = await structuralQueryEntities({
-        variables: {
-          query: {
-            filter: {
-              any: leftEntities.map((entity) => ({
-                equal: [
-                  { path: ["leftEntity", "uuid"] },
-                  {
-                    parameter: extractEntityUuidFromEntityId(
-                      entity.metadata.recordId.entityId,
-                    ),
-                  },
-                ],
-              })),
-            },
-            temporalAxes: currentTimeInstantTemporalAxes,
-            graphResolveDepths: {
-              ...zeroedGraphResolveDepths,
-              inheritsFrom: { outgoing: 255 },
-              isOfType: { outgoing: 2 },
-              hasRightEntity: { outgoing: 1, incoming: 0 },
-            },
-            includeDrafts: false,
-          },
-          includePermissions: false,
-        },
-      });
-
-      const outgoingLinksSubgraph = data
-        ? mapGqlSubgraphFieldsFragmentToSubgraph<EntityRootType>(
-            data.structuralQueryEntities.subgraph,
-          )
-        : undefined;
-
-      if (!outgoingLinksSubgraph) {
-        throw new Error("Could not fetch outgoing links of entities");
-      }
-
-      const outgoingLinkEntities = getRoots(
-        outgoingLinksSubgraph,
-      ) as LinkEntity[];
-
-      return outgoingLinkEntities.map((linkEntity) => {
-        const rightEntityRevisions = getRightEntityForLinkEntity(
-          outgoingLinksSubgraph,
-          linkEntity.metadata.recordId.entityId,
-        )!;
-
-        const rightEntity = rightEntityRevisions[0]!;
-
-        const rightEntityLabel = generateEntityLabel(
-          outgoingLinksSubgraph,
-          rightEntity,
-        );
-
-        const linkEntityType = getEntityTypeById(
-          outgoingLinksSubgraph,
-          linkEntity.metadata.entityTypeId,
-        )!;
-
-        return {
-          linkEntity,
-          rightEntity,
-          rightEntityLabel,
-          linkEntityType,
-        };
-      });
-    },
-    [structuralQueryEntities],
-  );
-
   const currentlyDisplayedRowsRef = useRef<TypeEntitiesRow[] | null>(null);
 
-  const generateCsvFile = useCallback(async () => {
-    const currentlyDisplayedRows = currentlyDisplayedRowsRef.current;
-    if (!currentlyDisplayedRows) {
-      return null;
-    }
-
-    // Table contents
-
-    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-    const columnRowKeys = columns.map(({ id }) => id ?? []).flat();
-
-    const tableContentColumnTitles = columns.map(({ title }) => title);
-
-    // Entity Properties
-
-    const propertyColumns = currentlyDisplayedRows.reduce<PropertyType[]>(
-      (prev, row) => {
-        const { entity } = row;
-
-        const propertyTypesUsedInEntity = Object.keys(entity.properties).map(
-          (baseUrl) => {
-            const propertyType = propertyTypes?.find(
-              ({ $id }) => extractBaseUrl($id) === baseUrl,
-            );
-
-            if (!propertyType) {
-              throw new Error(`Could not find property type for ${baseUrl}`);
-            }
-
-            return propertyType;
-          },
-        );
-
-        const newPropertyTypes = propertyTypesUsedInEntity.filter(
-          (propertyType) =>
-            !prev.some(
-              (previouslyAddedPropertyType) =>
-                previouslyAddedPropertyType.$id === propertyType.$id,
-            ),
-        );
-
-        return [...prev, ...newPropertyTypes];
-      },
-      [],
-    );
-
-    // Outgoing links
-
-    const outgoingLinksWithRightEntities = await fetchOutgoingLinksOfEntities({
-      leftEntities: currentlyDisplayedRows.map(({ entity }) => entity),
-    });
-
-    const outgoingLinkColumns = outgoingLinksWithRightEntities.reduce<
-      EntityType[]
-    >((prev, { linkEntityType }) => {
-      if (
-        !prev.some(
-          (previousLinkEntity) =>
-            previousLinkEntity.$id === linkEntityType.schema.$id,
-        )
-      ) {
-        return [...prev, linkEntityType.schema];
-      }
-
-      return prev;
-    }, []);
-
-    const content: string[][] = [
-      [
-        "Entity ID",
-        ...propertyColumns.map(({ title }) => title),
-        ...outgoingLinkColumns.map(({ title }) => title),
-        ...tableContentColumnTitles,
-      ],
-      ...currentlyDisplayedRows.map((row) => {
-        const { entity } = row;
-
-        const propertyValues = propertyColumns.map((propertyType) => {
-          /** @todo: stringify this better */
-          const propertyValue =
-            entity.properties[extractBaseUrl(propertyType.$id)];
-
-          if (typeof propertyValue === "string") {
-            return propertyValue;
-          } else if (typeof propertyValue === "object") {
-            return JSON.stringify(propertyValue);
-          } else if (propertyValue !== undefined) {
-            return String(propertyValue);
-          }
-
-          return "";
-        });
-
-        const outgoingLinks = outgoingLinksWithRightEntities.filter(
-          ({ linkEntity }) =>
-            linkEntity.linkData.leftEntityId ===
-            entity.metadata.recordId.entityId,
-        );
-
-        const outgoingLinkValues = outgoingLinkColumns.map((linkEntityType) => {
-          const outgoingLinksOfType = outgoingLinks.filter(
-            ({ linkEntityType: outgoingLinkEntityType }) =>
-              outgoingLinkEntityType.schema.$id === linkEntityType.$id,
-          );
-
-          if (outgoingLinksOfType.length > 0) {
-            return outgoingLinksOfType
-              .map(({ rightEntityLabel }) => rightEntityLabel)
-              .join(", ");
-          }
-
-          return "";
-        });
-
-        const tableContent = columnRowKeys.map((key) => {
-          const value = row[key];
-
-          if (typeof value === "string") {
-            return value;
-          } else if (key === "lastEditedBy") {
-            const user: MinimalUser | undefined = value;
-
-            return user?.preferredName ?? "";
-          }
-
-          return "";
-        });
-
-        return [
-          row.entityId,
-          ...propertyValues,
-          ...outgoingLinkValues,
-          ...tableContent,
-        ];
-      }),
-    ];
-
-    return {
-      title: "Entities",
-      content,
-    };
-  }, [
+  const { generateCsvFile } = useEntitiesTableGenerateCsvFile({
     currentlyDisplayedRowsRef,
     columns,
     propertyTypes,
-    fetchOutgoingLinksOfEntities,
-  ]);
+  });
 
   return (
     <Box>
